backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
boto3
python-dotenv
langchain
langchain-community
langchain-aws
pypdf2
python-docx
requests